requirements.txt
easydict==1.9
joblib==0.13.2
numpy==1.17.2
Pillow==6.1.0
pkg-resources==0.0.0
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
style==1.1.0
Theano==1.0.4
tornado==6.0.3
update==0.0.1
